+++ integration-tests/kyuubi-kubernetes-it/src/test/scala/org/apache/kyuubi/kubernetes/test/deployment/KyuubiOnKubernetesTestsSuite.scala
@@ -17,8 +17,8 @@
 
 package org.apache.kyuubi.kubernetes.test.deployment
 
-import org.apache.kyuubi.WithKyuubiServerOnKubernetes
 import org.apache.kyuubi.config.KyuubiConf
+import org.apache.kyuubi.kubernetes.test.WithKyuubiServerOnKubernetes
 import org.apache.kyuubi.operation.SparkQueryTests
 
 /**
